Refactor widgets to use MVVM and audio generation use case
