+++ app/src/main/java/com/jahi/pipelinetest/DailyCountdownWidget.kt
@@ -7,13 +7,9 @@ import android.appwidget.AppWidgetProvider
 import android.content.ComponentName
 import android.content.Context
 import android.content.Intent
-import android.media.MediaPlayer
 import android.widget.RemoteViews
-import java.io.File
-import java.net.HttpURLConnection
-import java.net.URL
-import kotlin.concurrent.thread
-import kotlin.random.Random
+import com.jahi.pipelinetest.domain.GenerateAudioUseCase
+import com.jahi.pipelinetest.viewmodel.WidgetViewModel
 import java.time.Duration
 import java.time.Instant
 import java.time.LocalDate
@@ -54,7 +50,8 @@ class DailyCountdownWidget : AppWidgetProvider() {
             }
             scheduleUpdates(context)
         } else if (intent.action == ACTION_GENERATE_AUDIO) {
-            generateAndPlay(context)
+            val vm = WidgetViewModel(GenerateAudioUseCase(context))
+            vm.playMotivationalAudio()
         }
     }
 
@@ -99,12 +96,6 @@ class DailyCountdownWidget : AppWidgetProvider() {
 
     companion object {
         const val ACTION_GENERATE_AUDIO = "com.jahi.pipelinetest.GENERATE_AUDIO"
-
-        private val motivationalSentences = listOf(
-            "You can achieve anything you set your mind to.",
-            "Believe in yourself and all that you are.",
-            "Every day is a chance to get better."
-        )
 
         fun updateAppWidget(context: Context, appWidgetManager: AppWidgetManager, appWidgetId: Int) {
             val views = RemoteViews(context.packageName, R.layout.daily_countdown_widget)
@@ -155,46 +146,5 @@ class DailyCountdownWidget : AppWidgetProvider() {
             }
         }
 
-        private fun generateAndPlay(context: Context) {
-            thread {
-                val text = motivationalSentences[Random.nextInt(motivationalSentences.size)]
-                val file = fetchAudio(context, text)
-                file?.let { playAudio(it) }
-            }
-        }
-
-        private fun fetchAudio(context: Context, text: String): File? {
-            return try {
-                val url = URL("https://api.elevenlabs.io/v1/text-to-speech/EXAVITQu4vr4xnSDxMaL/stream")
-                val conn = url.openConnection() as HttpURLConnection
-                conn.requestMethod = "POST"
-                conn.setRequestProperty("xi-api-key", BuildConfig.ELEVEN_LAB_KEY)
-                conn.setRequestProperty("Content-Type", "application/json")
-                conn.doOutput = true
-                val body = "{\"text\": \"$text\"}"
-                conn.outputStream.use { it.write(body.toByteArray()) }
-                if (conn.responseCode == HttpURLConnection.HTTP_OK) {
-                    val file = File.createTempFile("tts", ".mp3", context.cacheDir)
-                    conn.inputStream.use { input -> file.outputStream().use { input.copyTo(it) } }
-                    file
-                } else null
-            } catch (e: Exception) {
-                null
-            }
-        }
-
-        private fun playAudio(file: File) {
-            try {
-                val mp = MediaPlayer()
-                mp.setDataSource(file.absolutePath)
-                mp.setOnCompletionListener {
-                    it.release()
-                    file.delete()
-                }
-                mp.prepare()
-                mp.start()
-            } catch (_: Exception) {
-            }
-        }
     }
 }
