+++ app/src/main/java/com/jahi/pipelinetest/viewmodel/WidgetViewModel.kt
@@ -1,0 +1,13 @@
+package com.jahi.pipelinetest.viewmodel
+
+import androidx.lifecycle.ViewModel
+import com.jahi.pipelinetest.domain.GenerateAudioUseCase
+
+class WidgetViewModel(
+    private val generateAudioUseCase: GenerateAudioUseCase
+) : ViewModel() {
+
+    fun playMotivationalAudio() {
+        generateAudioUseCase()
+    }
+}
